rename types -> errors
